More docstrings.
Added _queries.py for better organization.
