+++ tests/test_other_mixins.py
@@ -1,6 +1,3 @@
-import mock
-import pytest
-
 from django.contrib import messages
 from django.contrib.messages.middleware import MessageMiddleware
 from django.contrib.messages.storage.base import Message
@@ -18,52 +15,42 @@ from braces.views import (
     FormInvalidMessageMixin,
 )
 from .factories import UserFactory
-from .helpers import TestViewHelper
 from .models import Article, CanonicalArticle
 from .views import (
-    ArticleListView,
-    ArticleListViewWithCustomQueryset,
-    AuthorDetailView,
-    OrderableListView,
     FormMessagesView,
     ContextView,
 )
 
 
 class TestSuccessURLRedirectListMixin(test.TestCase):
-    """
-    Tests for SuccessURLRedirectListMixin.
-    """
+    """Scenarios around redirecting after a successful form submission"""
 
     def test_redirect(self):
-        """
-        Test if browser is redirected to list view.
-        """
+        """Successful POST should redirect"""
         data = {"title": "Test body", "body": "Test body"}
         resp = self.client.post("/article_list/create/", data)
         self.assertRedirects(resp, "/article_list/")
 
     def test_no_url_name(self):
-        """
-        Test that ImproperlyConfigured is raised.
-        """
+        """Improper setup should raise ImproperlyConfigured"""
         data = {"title": "Test body", "body": "Test body"}
         with self.assertRaises(ImproperlyConfigured):
+            # The view at this endpoint has no success_url defined
             self.client.post("/article_list_bad/create/", data)
 
 
 class TestUserFormKwargsMixin(test.TestCase):
-    """
-    Tests for UserFormKwargsMixin.
-    """
+    """Scenarios around automatically including a user in form submissions"""
 
     def test_post_method(self):
+        """A POST request should include the user kwarg"""
         user = UserFactory()
         self.client.login(username=user.username, password="asdf1234")
         resp = self.client.post("/form_with_user_kwarg/", {"field1": "foo"})
-        assert force_str(resp.content) == "username: %s" % user.username
+        assert force_str(resp.content) == f"username: {user.username}"
 
     def test_get_method(self):
+        """A GET request should include the user kwarg"""
         user = UserFactory()
         self.client.login(username=user.username, password="asdf1234")
         resp = self.client.get("/form_with_user_kwarg/")
@@ -71,53 +58,45 @@ class TestUserFormKwargsMixin(test.TestCase):
 
 
 class TestSetHeadlineMixin(test.TestCase):
-    """
-    Tests for SetHeadlineMixin.
-    """
+    """Scenarios around setting a headline"""
 
     def test_dynamic_headline(self):
-        """
-        Tests if get_headline() is called properly.
-        """
+        """A method-provided headline should be included in context"""
         resp = self.client.get("/headline/test-headline/")
         self.assertEqual("test-headline", resp.context["headline"])
 
     def test_context_data(self):
-        """
-        Tests if mixin adds proper headline to template context.
-        """
+        """An attribute-provided headline should be in the context"""
         resp = self.client.get("/headline/foo-bar/")
         self.assertEqual("foo-bar", resp.context["headline"])
 
-    def test_get_headline(self):
-        """
-        Tests if get_headline() method works correctly.
-        """
+    def test_improper_configuration(self):
+        """Not providing a headline should raise an exception"""
         mixin = SetHeadlineMixin()
         with self.assertRaises(ImproperlyConfigured):
             mixin.get_headline()
 
-        mixin.headline = "Test headline"
-        self.assertEqual("Test headline", mixin.get_headline())
-
     def test_get_headline_lazy(self):
+        """Lazy evaluation of text should still provide the headline"""
         resp = self.client.get("/headline/lazy/")
         self.assertEqual("Test Headline", resp.context["headline"])
 
 
 class TestStaticContextMixin(test.TestCase):
-    """Tests for StaticContextMixin."""
+    """Scenarios around including static content in the context"""
 
     view_class = ContextView
     view_url = "/context/"
 
-    def test_dict(self):
+    def test_dictionary(self):
+        """Static content can be included as a dictionary"""
         self.view_class.static_context = {"test": True}
         resp = self.client.get(self.view_url)
         self.assertEqual(200, resp.status_code)
         self.assertEqual(True, resp.context["test"])
 
     def test_two_tuple(self):
+        """Static content can be included as a two-tuple pair"""
         self.view_class.static_context = [("a", 1), ("b", 2)]
         resp = self.client.get(self.view_url)
         self.assertEqual(200, resp.status_code)
@@ -125,307 +104,43 @@ class TestStaticContextMixin(test.TestCase):
         self.assertEqual(2, resp.context["b"])
 
     def test_not_set(self):
+        """ImproperlyConfigured should be raised if no static content is set"""
         self.view_class.static_context = None
         with self.assertRaises(ImproperlyConfigured):
             self.client.get(self.view_url)
 
     def test_string_value_error(self):
+        """A string should raise ImproperlyConfigured"""
         self.view_class.static_context = "Fail"
         with self.assertRaises(ImproperlyConfigured):
             self.client.get(self.view_url)
 
     def test_list_error(self):
+        """A list should raise ImproperlyConfigured"""
         self.view_class.static_context = ["fail", "fail"]
         with self.assertRaises(ImproperlyConfigured):
             self.client.get(self.view_url)
 
 
 class TestCsrfExemptMixin(test.TestCase):
-    """
-    Tests for TestCsrfExemptMixin.
-    """
+    """Scenarios around views which are CSRF exempt"""
 
     def setUp(self):
+        """Ensure the client enforces CSRF checks"""
         super(TestCsrfExemptMixin, self).setUp()
         self.client = self.client_class(enforce_csrf_checks=True)
 
     def test_csrf_token_is_not_required(self):
-        """
-        Tests if csrf token is not required.
-        """
+        """CSRF tokens should not be required"""
         resp = self.client.post("/csrf_exempt/", {"field1": "test"})
         self.assertEqual(200, resp.status_code)
         self.assertEqual("OK", force_str(resp.content))
 
 
-class TestSelectRelatedMixin(TestViewHelper, test.TestCase):
-    view_class = ArticleListView
-
-    def test_missing_select_related(self):
-        """
-        ImproperlyConfigured exception should be raised if select_related
-        attribute is missing.
-        """
-        with self.assertRaises(ImproperlyConfigured):
-            self.dispatch_view(self.build_request(), select_related=None)
-
-    def test_invalid_select_related(self):
-        """
-        ImproperlyConfigured exception should be raised if select_related is
-        not a tuple or a list.
-        :return:
-        """
-        with self.assertRaises(ImproperlyConfigured):
-            self.dispatch_view(self.build_request(), select_related={"a": 1})
-
-    @mock.patch("django.db.models.query.QuerySet.select_related")
-    def test_select_related_called(self, m):
-        """
-        Checks if QuerySet's select_related() was called with correct
-        arguments.
-        """
-        qs = Article.objects.all()
-        m.return_value = qs.select_related("author")
-        qs.select_related = m
-        m.reset_mock()
-
-        resp = self.dispatch_view(self.build_request())
-        self.assertEqual(200, resp.status_code)
-        m.assert_called_once_with("author")
-
-    @mock.patch("django.db.models.query.QuerySet.select_related")
-    def test_select_related_keeps_select_related_from_queryset(self, m):
-        """
-        Checks that an empty select_related attribute does not
-        cancel a select_related provided by queryset.
-        """
-        qs = Article.objects.all()
-        qs.select_related = m
-        m.reset_mock()
-
-        with pytest.warns(UserWarning):
-            resp = self.dispatch_view(
-                self.build_request(),
-                view_class=ArticleListViewWithCustomQueryset,
-            )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(0, m.call_count)
-
-
-class TestPrefetchRelatedMixin(TestViewHelper, test.TestCase):
-    view_class = AuthorDetailView
-
-    def test_missing_prefetch_related(self):
-        """
-        ImproperlyConfigured exception should be raised if
-        prefetch_related attribute is missing.
-        """
-        with self.assertRaises(ImproperlyConfigured):
-            self.dispatch_view(self.build_request(), prefetch_related=None)
-
-    def test_invalid_prefetch_related(self):
-        """
-        ImproperlyConfigured exception should be raised if
-        prefetch_related is not a tuple or a list.
-        :return:
-        """
-        with self.assertRaises(ImproperlyConfigured):
-            self.dispatch_view(self.build_request(), prefetch_related={"a": 1})
-
-    @mock.patch("django.db.models.query.QuerySet.prefetch_related")
-    def test_prefetch_related_called(self, m):
-        """
-        Checks if QuerySet's prefetch_related() was called with correct
-        arguments.
-        """
-        qs = Article.objects.all()
-        m.return_value = qs.prefetch_related("article_set")
-        qs.prefetch_related = m
-        m.reset_mock()
-
-        resp = self.dispatch_view(self.build_request())
-        self.assertEqual(200, resp.status_code)
-        m.assert_called_once_with("article_set")
-
-    @mock.patch("django.db.models.query.QuerySet.prefetch_related")
-    def test_prefetch_related_keeps_select_related_from_queryset(self, m):
-        """
-        Checks that an empty prefetch_related attribute does not
-        cancel a prefetch_related provided by queryset.
-        """
-        qs = Article.objects.all()
-        qs.prefetch_related = m
-        m.reset_mock()
-
-        with pytest.warns(UserWarning):
-            resp = self.dispatch_view(
-                self.build_request(),
-                view_class=ArticleListViewWithCustomQueryset,
-            )
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(0, m.call_count)
-
-
-class TestOrderableListMixin(TestViewHelper, test.TestCase):
-    view_class = OrderableListView
-
-    def __make_test_articles(self):
-        a1 = Article.objects.create(title="Alpha", body="Zet")
-        a2 = Article.objects.create(title="Zet", body="Alpha")
-        return a1, a2
-
-    def test_correct_order(self):
-        """
-        Objects must be properly ordered if requested with valid column names
-        """
-        a1, a2 = self.__make_test_articles()
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=title&ordering=asc"),
-            orderable_columns=None,
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=id&ordering=desc"),
-            orderable_columns=None,
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a2, a1])
-
-    def test_correct_order_with_default_ordering(self):
-        """
-        Objects must be properly ordered if requested with valid column names
-        and with the default ordering
-        """
-        a1, a2 = self.__make_test_articles()
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=id"),
-            orderable_columns=None,
-            ordering_default=None,
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=id"),
-            orderable_columns=None,
-            ordering_default="asc",
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=id"),
-            orderable_columns=None,
-            ordering_default="desc",
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a2, a1])
-
-    def test_correct_order_with_param_not_default_ordering(self):
-        """
-        Objects must be properly ordered if requested with valid column names
-        and ordering option in the query params.
-        In this case, the ordering_default will be overwritten.
-        """
-        a1, a2 = self.__make_test_articles()
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=id&ordering=asc"),
-            orderable_columns=None,
-            ordering_default="desc",
-            get_orderable_columns=lambda: (
-                "id",
-                "title",
-            ),
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-    def test_correct_order_with_incorrect_default_ordering(self):
-        """
-        Objects must be properly ordered if requested with valid column names
-        and with the default ordering
-        """
-        view = self.view_class()
-        view.ordering_default = "improper_default_value"
-        self.assertRaises(
-            ImproperlyConfigured, lambda: view.get_ordering_default()
-        )
-
-    def test_default_column(self):
-        """
-        When no ordering specified in GET, use
-        View.get_orderable_columns_default()
-        """
-        a1, a2 = self.__make_test_articles()
-
-        resp = self.dispatch_view(self.build_request())
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-    def test_get_orderable_columns_returns_correct_values(self):
-        """
-        OrderableListMixin.get_orderable_columns() should return
-        View.orderable_columns attribute by default or raise
-        ImproperlyConfigured exception in the attribute is None
-        """
-        view = self.view_class()
-        self.assertEqual(view.get_orderable_columns(), view.orderable_columns)
-        view.orderable_columns = None
-        self.assertRaises(
-            ImproperlyConfigured, lambda: view.get_orderable_columns()
-        )
-
-    def test_get_orderable_columns_default_returns_correct_values(self):
-        """
-        OrderableListMixin.get_orderable_columns_default() should return
-        View.orderable_columns_default attribute by default or raise
-        ImproperlyConfigured exception in the attribute is None
-        """
-        view = self.view_class()
-        self.assertEqual(
-            view.get_orderable_columns_default(),
-            view.orderable_columns_default,
-        )
-        view.orderable_columns_default = None
-        self.assertRaises(
-            ImproperlyConfigured, lambda: view.get_orderable_columns_default()
-        )
-
-    def test_only_allowed_columns(self):
-        """
-        If column is not in Model.Orderable.columns iterable, the objects
-        should be ordered by default column.
-        """
-        a1, a2 = self.__make_test_articles()
-
-        resp = self.dispatch_view(
-            self.build_request(path="?order_by=body&ordering=asc"),
-            orderable_columns_default=None,
-            get_orderable_columns_default=lambda: "title",
-        )
-        self.assertEqual(list(resp.context_data["object_list"]), [a1, a2])
-
-
 class TestCanonicalSlugDetailView(test.TestCase):
+    """Scenarios involving canonical slugs"""
     def setUp(self):
+        """Create the two articles"""
         Article.objects.create(title="Alpha", body="Zet", slug="alpha")
         Article.objects.create(title="Zet", body="Alpha", slug="zet")
 
@@ -449,7 +164,9 @@ class TestCanonicalSlugDetailView(test.TestCase):
 
 
 class TestNamespaceAwareCanonicalSlugDetailView(test.TestCase):
+    """Scenarios around canonical slugs and namespaces"""
     def setUp(self):
+        """Create the necessary articles"""
         Article.objects.create(title="Alpha", body="Zet", slug="alpha")
         Article.objects.create(title="Zet", body="Alpha", slug="zet")
 
@@ -480,7 +197,9 @@ class TestNamespaceAwareCanonicalSlugDetailView(test.TestCase):
 
 
 class TestOverriddenCanonicalSlugDetailView(test.TestCase):
+    """Scenarios involving overridden canonical slugs"""
     def setUp(self):
+        """Create the necessary articles"""
         Article.objects.create(title="Alpha", body="Zet", slug="alpha")
         Article.objects.create(title="Zet", body="Alpha", slug="zet")
 
@@ -505,7 +224,9 @@ class TestOverriddenCanonicalSlugDetailView(test.TestCase):
 
 
 class TestCustomUrlKwargsCanonicalSlugDetailView(test.TestCase):
+    """Scenarios around canonical slugs and custom URL kwargs"""
     def setUp(self):
+        """Create the articles"""
         Article.objects.create(title="Alpha", body="Zet", slug="alpha")
         Article.objects.create(title="Zet", body="Alpha", slug="zet")
 
@@ -529,7 +250,9 @@ class TestCustomUrlKwargsCanonicalSlugDetailView(test.TestCase):
 
 
 class TestModelCanonicalSlugDetailView(test.TestCase):
+    """Scenarios around canonical slugs and model fields"""
     def setUp(self):
+        """Generate the necessary articles"""
         CanonicalArticle.objects.create(
             title="Alpha", body="Zet", slug="alpha"
         )
@@ -560,26 +283,32 @@ class TestModelCanonicalSlugDetailView(test.TestCase):
     MESSAGE_STORAGE="django.contrib.messages.storage.cookie.CookieStorage"
 )
 class MessageMixinTests(test.TestCase):
+    """Scenarios around the messaging framework"""
     def setUp(self):
+        """Create necessary objects"""
         self.rf = test.RequestFactory()
         self.middleware = MessageMiddleware("")
 
     def get_request(self, *args, **kwargs):
+        """Generate a request that has passed through the middleware"""
         request = self.rf.get("/")
         self.middleware.process_request(request)
         return request
 
     def get_response(self, request, view):
+        """Generate a response that has been passed through the middleware"""
         response = view(request)
         self.middleware.process_response(request, response)
         return response
 
     def get_request_response(self, view, *args, **kwargs):
+        """Get both a request and a response, middleware-processed"""
         request = self.get_request(*args, **kwargs)
         response = self.get_response(request, view)
         return request, response
 
     def test_add_messages(self):
+        """Message should be added through the class attribute"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.add_message(messages.SUCCESS, "test")
@@ -592,6 +321,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0].level, messages.SUCCESS)
 
     def test_get_messages(self):
+        """get_messages should get the stored messages"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.add_message(messages.SUCCESS, "success")
@@ -605,6 +335,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(response.content, b"success,warning")
 
     def test_get_level(self):
+        """Should be able to get message levels"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 return HttpResponse(self.messages.get_level())
@@ -613,6 +344,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(int(response.content), messages.INFO)  # default
 
     def test_set_level(self):
+        """Should be able to set message levels"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.set_level(messages.WARNING)
@@ -626,6 +358,7 @@ class MessageMixinTests(test.TestCase):
 
     @override_settings(MESSAGE_LEVEL=messages.DEBUG)
     def test_debug(self):
+        """Messages should able to be set as DEBUG"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.debug("test")
@@ -637,6 +370,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0], Message(messages.DEBUG, "test"))
 
     def test_info(self):
+        """Messages should able to be set as INFO"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.info("test")
@@ -648,6 +382,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0], Message(messages.INFO, "test"))
 
     def test_success(self):
+        """Messages should able to be set as SUCCESS"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.success("test")
@@ -659,6 +394,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0], Message(messages.SUCCESS, "test"))
 
     def test_warning(self):
+        """Messages should able to be set as WARNING"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.warning("test")
@@ -670,6 +406,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0], Message(messages.WARNING, "test"))
 
     def test_error(self):
+        """Messages should able to be set as ERROR"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.error("test")
@@ -681,6 +418,7 @@ class MessageMixinTests(test.TestCase):
         self.assertEqual(msg[0], Message(messages.ERROR, "test"))
 
     def test_invalid_attribute(self):
+        """Raise an AttributeError if setting an invalid level"""
         class TestView(MessageMixin, View):
             def get(self, request):
                 self.messages.invalid()
@@ -720,11 +458,13 @@ class MessageMixinTests(test.TestCase):
 
 
 class TestFormMessageMixins(test.TestCase):
+    """Scenarios around form valid/invalid messages"""
     def setUp(self):
         self.good_data = {"title": "Good", "body": "Body"}
         self.bad_data = {"body": "Missing title"}
 
     def test_valid_message(self):
+        """If the form is valid, the valid message should be available"""
         url = "/form_messages/"
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
@@ -734,6 +474,7 @@ class TestFormMessageMixins(test.TestCase):
         self.assertContains(response, FormMessagesView().form_valid_message)
 
     def test_invalid_message(self):
+        """If the form is invalid, the invalid message should be available"""
         url = "/form_messages/"
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
@@ -743,96 +484,123 @@ class TestFormMessageMixins(test.TestCase):
         self.assertContains(response, FormMessagesView().form_invalid_message)
 
     def test_form_valid_message_not_set(self):
+        """Not setting a form_valid message should raise ImproperlyConfigured"""
         mixin = FormValidMessageMixin()
         with self.assertRaises(ImproperlyConfigured):
             mixin.get_form_valid_message()
 
     def test_form_valid_message_not_str(self):
+        """Non-strings for the form_valid message should raise ImproperlyConfigured"""
         mixin = FormValidMessageMixin()
         mixin.form_valid_message = ["bad"]
         with self.assertRaises(ImproperlyConfigured):
             mixin.get_form_valid_message()
 
     def test_form_valid_returns_message(self):
+        """get_form_valid_message should return the form_valid message"""
         mixin = FormValidMessageMixin()
         mixin.form_valid_message = "Good øø"
         self.assertEqual(force_str("Good øø"), mixin.get_form_valid_message())
 
     def test_form_invalid_message_not_set(self):
+        """Not setting a form_invalid message should raise ImproperlyConfigured"""
         mixin = FormInvalidMessageMixin()
         with self.assertRaises(ImproperlyConfigured):
             mixin.get_form_invalid_message()
 
     def test_form_invalid_message_not_str(self):
+        """Non-strings for the form_invalid message should raise ImproperlyConfigured"""
         mixin = FormInvalidMessageMixin()
         mixin.form_invalid_message = ["bad"]
         with self.assertRaises(ImproperlyConfigured):
             mixin.get_form_invalid_message()
 
     def test_form_invalid_returns_message(self):
+        """get_form_invalid_message should return the form_invalid message"""
         mixin = FormInvalidMessageMixin()
         mixin.form_invalid_message = "Bad øø"
         self.assertEqual(force_str("Bad øø"), mixin.get_form_invalid_message())
 
 
 class TestAllVerbsMixin(test.TestCase):
+    """Scenarios around the AllVerbsMixin"""
     def setUp(self):
         self.url = "/all_verbs/"
         self.no_handler_url = "/all_verbs_no_handler/"
 
     def test_options(self):
+        """AllVerbs should respond to OPTION"""
         response = self.client.options(self.url)
         self.assertEqual(response.status_code, 200)
 
     def test_get(self):
+        """AllVerbs should respond to GET"""
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, 200)
 
     def test_head(self):
+        """AllVerbs should respond to HEAD"""
         response = self.client.head(self.url)
         self.assertEqual(response.status_code, 200)
 
     def test_post(self):
+        """AllVerbs should respond to POST"""
         response = self.client.post(self.url)
         self.assertEqual(response.status_code, 200)
 
     def test_put(self):
+        """AllVerbs should respond to PUT"""
         response = self.client.put(self.url)
         self.assertEqual(response.status_code, 200)
 
     def test_delete(self):
+        """AllVerbs should respond to DELETE"""
         response = self.client.delete(self.url)
         self.assertEqual(response.status_code, 200)
 
+    def test_patch(self):
+        """AllVerbs should respond to PATCH"""
+        response = self.client.patch(self.url)
+        self.assertEqual(response.status_code, 200)
+
     def test_no_all_handler(self):
+        """A missing handler should raise ImproperlyConfigured"""
         with self.assertRaises(ImproperlyConfigured):
             self.client.get("/all_verbs_no_handler/")
 
 
 class TestHeaderMixin(test.TestCase):
+    """Scenarios around the extra headers mixin"""
     def test_attribute(self):
+        """Headers can be set via an attribute"""
         response = self.client.get("/headers/attribute/")
         self.assertEqual(response["X-DJANGO-BRACES-1"], "1")
         self.assertEqual(response["X-DJANGO-BRACES-2"], "2")
 
     def test_method(self):
+        """Headers can be set via a method"""
         response = self.client.get("/headers/method/")
         self.assertEqual(response["X-DJANGO-BRACES-1"], "1")
         self.assertEqual(response["X-DJANGO-BRACES-2"], "2")
 
     def test_existing(self):
+        """Existing headers should still come through"""
         response = self.client.get('/headers/existing/')
         self.assertEqual(response['X-DJANGO-BRACES-EXISTING'], 'value')
 
 class TestCacheControlMixin(test.TestCase):
+    """Scenarios around controlling cache"""
     def test_cachecontrol_public(self):
+        """Cache settings should be respected and included"""
         response = self.client.get('/cachecontrol/public/')
         options = [i.strip() for i in response['Cache-Control'].split(',')]
         self.assertEqual(sorted(options), ['max-age=60', 'public'])
 
 
 class TestNeverCacheMixin(test.TestCase):
+    """Scenarios around marking a view as never-cached"""
     def test_nevercache(self):
+        """Views marked as no-cache should not be cached"""
         response = self.client.get('/nevercache/')
         options = [i.strip() for i in response['Cache-Control'].split(',')]
         expected_cache_control_options = {"max-age=0", "must-revalidate", "no-cache", "no-store", "private"}
